refactoring kzg
